refactor(context_compressor): extract threshold coercion mixin (LB7)
Window 1307-1422 (class attrs + 5 static/instance methods) into
ContextCompressorThresholdMixin (MRO seam, SessionTitleMixin precedent).
Mixin composed into ContextCompressor bases; members resolve is-identical
through the class. One class-name reference retargeted to the mixin.
Deps one-way on budget module + model_metadata. 52/52 targeted tests green.

Part of #78645
Part of #78647
